JVMCBC-298: Implement 0x97 to get last closed checkpoint
Motivation
----------
DCP stream continuation requires the point in history of partition. One
way is to persist it on the consumer side. But to implement "listen from
now" strategy there is should be a way to get the last sequence number.

Modification
------------
Implement handler for command 0x97 which returns sequence number for
given partition.

Result
------
It is possible to fetch sequence number for arbitrary partition, and use
it to initialize DCP stream.

Change-Id: Id203630f1beb59f54e18ae8c3de63779c2ae4545
Reviewed-on: http://review.couchbase.org/60731
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>