[cleanup] Reformatting and Refactoring of RequestHandler
Motivation
----------
This changeset slightly reformats the RequestHandler and also splits out
the big "onEvent" method into smaller method chunks. This allows for
better inlining and easier maintenance.

The only functional change is that when a new config arrives it is
always published onto the event bus, since the bus itself checks if
it has subscribers (avoid double checking).

The branching on found Nodes has been improved to have the common case
(nodes found) in the if block.

Change-Id: I25da18c8691aea939b855792c4f149c33df56116
Reviewed-on: http://review.couchbase.org/60830
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>