JVMCBC-305: Allow DCP connectors to notify core about processing messages
Motivation
----------
To implement flow control efficiently, it is necessary to be able to
receive notification that the consumer really processed the
event (transformed, persisted or retransmitted)

Modification
------------
Persist network body size in the DCP message class, expose connection
reference, and the API to notify connection that particular message has
been processed in upstream. DCPConnection tracks the amount of data
processed and issue acknowledgment packets for flow control on the server.

Result
------
Connectors will not suffer of OOM, when internal buffer are not
capable to handle high speed DCP traffic.

Change-Id: I4898d4632eb1d8b80f526e6650241977e9b681c5
Reviewed-on: http://review.couchbase.org/60831
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>