JVMCBC-306: Allow optional Observable completion on the Event Loops
Motiviation
-----------
As an advanced option it should be possible to complete the observables
on the IO event loops. This is disabled by default for safety reasons,
but if the user knows what they're doing they can get noticeable
improvements because we need to do less movement across thread
boundaries.

Modifications
-------------
The appropriate environment option has been added and the functionality
implemented in the netty handler.

The builder has a warning sign on it saying this is only for advanced
usage.

Result
------
Potentially better performance for users who know what they are
doing.

Change-Id: I1780a9f21664b00b32a75475e49dcd1f3a0705a0
Reviewed-on: http://review.couchbase.org/61170
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>