JVMCBC-310: Do not encode key twice.
Motivation
----------
Right now the netty memcached codec takes the key as a string
and performs encoding internally. This is problematic, since
the SDK itself also needs to encode the key for routing
purposes so as a result the key ends up to be encoded twice
from its String representation into a UTF-8 byte array.

This produces unnecessary garbage and wastes CPU cycles.

Modifications
-------------
We already encode the key into a byte array when the request
is created, so the netty codec is changed to take a byte[]
instead of a String and write it down directly.

Note that while upstream netty has a change implemented that
makes the key a ByteBuf, this intermediate change has nearly
the same effect and allows us to reduce the leak potential
to a minimum. Once internally we upgrade to netty 4.1, this
needs to be revisited since it will break the API (internally).

This changeset has no visible API or behavior impact for the user.

Result
------
Less garbage created and less work to do on the hot code path.

Change-Id: I7d896108b7a22bda54e1ff6dbfda431a4e038cd7
Reviewed-on: http://review.couchbase.org/61414
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>