JVMCBC-309: Gracefully handle errors during metrics collection.
Motivation
----------
Since the underlying histogram can have bugs or we feed it invalid
timings somehow, it can throw an exception and as a result it will
bubble into our IO event loops.

Since metrics collection is considered support functionality, it
should under no circumstance interfer with the actual IO processing.

Modifications
-------------
try/catch is placed around the metrics collection logic and if an
error is detected it is logged for proper inspection.

Result
------
If errors happen during latency metrics collection, they are not
affecting the actual IO workload  but are logged for proper
supportability.

Change-Id: I903f7b6d9c236c5cd3b727513bbd39cb4459a974
Reviewed-on: http://review.couchbase.org/61413
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>