JVMCBC-317: Take Netty 4.0.35 bundled epoll into account
Since Netty 4.0.35, the native epoll library comes bundled with the
artifact, and a property can be used to indicate that Netty is shaded.

This change sets up the system property to indicate to Netty that its
classes have been shaded and relocated.

It also modifies the pom so that Netty newly integrated epoll native
library is integrated in the shaded jar as a renamed/relocated version.

Due to the way the Shade plugin transformer works, this implies:
 - extracting the .so from the dependency itself, using the dependency
   plugin, during prepare-package.
 - using a transformer in the Shade plugin to rename that extracted lib.
 - using a filter to ensure the Shade plugin doesn't also include the
   original lib.

This is the most straightforward way we found to achieve this goal while
not being to intrusive on the IDE (eg. avoid IntelliJ picking the
unpacked .so as a resource folder).

Change-Id: I633846ece7b18eb0bd3bc866f0737d0e6d66ed91
Reviewed-on: http://review.couchbase.org/61986
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>