JVMCBC-320: Lower the number of IO event loops by default to 3.
Motivation
----------
The default number of the "ioPoolSize" has always been the number
of processors returned by the JVM - This is the default as well in
netty.

In several benchmarking and profiling scenarios there has been no
evidence found that a larger number of event loops is better, since
we are opening hundreds of connections max. Netty has been designed
to handle thousands of those at the same time.

Modifications
-------------
As a result, turning the number of event loops down to 3 by default,
while still keeping it configurable should have no negative impact
on workloads, even in high throughput scenarios.

In fact, the upcoming changes in netty which allow to plug in selector
strategies work better on smaller number of event loops since the cpu
resources are divided with less context switching needed. Other effects
like selector wakeups are also improved since less of them are needed
and traffic is condensed on those threads.

Result
------
Better overall traffic distribution on less number of event loops by
default, since we are running a client and not a server which needs to
handle thousands of connections all the time.

Change-Id: Icf7892f73433c30cef90aa4dcc4531e6456e3692
Reviewed-on: http://review.couchbase.org/62235
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>