JVMCBC-323: Gracefully handle config completion errors during shutdown.
Motivation
----------
When the handler is removed during shutdown, it might be the case that
the scheduler where this is run on is already shutting down and as a
result not taking any more tasks.

Modifications
-------------
Catch the specific exception and print a log message, but don't let it
bubble through the event loop. It's not harmful, but looks a little
nasty in the logs.

Result
------
Less clutter in the logs when this happens during (very often violent)
shutdown.

Change-Id: I64bba849bb000110359f6f49641d4395a0d34442
Reviewed-on: http://review.couchbase.org/62294
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>