JVMCBC-324: Allow some requests (n1ql) to target specific node
Motivation:
-----------
The need has come up for some N1QL queries to be able to target specific
nodes, in order to emulate a broadcast or multicast on some operations.

Modifications:
--------------
Introduced thePrelocatedRequest interface that can be used to mark such
requests (and has a getter for the address on which to dispatch the
request).

Only QueryLocator takes that interface into account (for now), and any
GenericQueryRequest could be targeted at a specific node that way.

Results:
--------
GenericQueryRequests can be targeted at specific nodes instead of
delegating to QueryLocator strategy.

Change-Id: Iae6292429f876ed0bac333fcedaaf40c35b49c3c
Reviewed-on: http://review.couchbase.org/62346
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>