JVMCBC-325: Send Host header on all outgoing HTTP requests.
Motivation
----------
go 1.6 recently started to check for the mandatory Host header (according
to the RFC). This was not an issue in the past since no component on the
server did care, but now it does. Only the FTS service is affected for now
but when query upgrades this might become an issue there too.

Modifications
-------------
Send the Host header, but cache it in the context so that repeated use
doesn't have to call toString() every time.

Result
------
Compat with go 1.6 on the server side and in general with the HTTP
RFC.

Change-Id: Iff8e2528dba549d22e963544b08977983087fbfb
Reviewed-on: http://review.couchbase.org/62831
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>