JVMCBC-326: Fix parsing of json strings ending with 2 backslashes
Motivation:
-----------
A customer reported that a N1QL query was looping until timeout with
their test dataset.

When looking for row boundaries, both view and query parsers use
ClosingPositionBufProcessor to skip overs what it detects as being part
of a JSON string. To do so, it tracks the end of strings by looking for
a closing quote and checking that it is not preceded by an escaping
backslash.

The thing is, such a backslash could itself be escaped inside the string
by another backslash, and that later case isn't detected, wrongly
assuming the inspected quote isn't the closing one.

Modifications:
--------------
Both ClosingPositionBufProcessor and StringClosingPositionBufProcessor
have json string boundaries detection fixed by remembering the last
processed character AND the one before it. This allows to detect that a
quote preceded by two backslashes is a closing quote.

Added corresponding unit tests to the processors tests.

Added QueryHandler unit test with a payload simulating the same error
conditions as those the customer encountered.

Results:
--------
Streaming parsers can now correctly detect string boundaries of json
strings that end with two backslashes.

QueryHandler can correctly process a payload with double backslashes at
the end of a json string.

Change-Id: I9e3408832aa062a3a9244b02d6a7ea5d07fd9abd
Reviewed-on: http://review.couchbase.org/62919
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>