JVMCBC-327: Fix SocketAddress cast in generic handler
Motivation
----------
The AbstractGenericHandler remoteHttpHost retrieves the SocketAddress
from the ChannelHandlerContext but it performs an unverified cast to
a InetSocketAddress. In some instances, most notably tests with an
EmbeddedChannel, the cast can fail.

Modifications
-------------
Added an instanceOf check before the cast, and using toString as a
fallback mechanism (which should only occur in testing situations
anyway, since when the remoteHttpHost method matters is when we are
communicating with a REST endpoint, so through a InetSocketAddress).

Result
------
No more unit tests failing due to a ClassCastException.

Change-Id: I6f1db47f028260779376367e6a0dabdc7dfc468b
Reviewed-on: http://review.couchbase.org/62949
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
(backported via cherry pick from commit 0491580)
Reviewed-on: http://review.couchbase.org/63470