JVMCBC-330: Enable trace logging of CouchbaseRequests on retry
Motivation
----------
Since a retry of an operation (especially in a steady state) is
something which should be unusual, it makes sense to log the
amount and retry delay of those requests at TRACE level (to not
spam the logs in an overload condition).

This also allows, if there is a suspicion of retry events going
on but they shouldn't, to specifically enable TRACE logging for
the ResponseHandler only.

Modifications
-------------
The ResponseHandler now logs the retry event with the actual request
as well as the retry delay and unit used.

Result
------
More insight into how and when requests are retried.

Change-Id: I74c5492a4c92123930c987c6bd3690c7fe39f9b3
Reviewed-on: http://review.couchbase.org/63402
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>