JVMCBC-331: Correctly shut down PauseDetector when used more than once.
Motivation
----------
Right now we only use one latency metrics collector, but if we use more
than one in the future the shutdown of the pause detector would be called
more than once on the one static object.

Modifications
-------------
Instead of shutting down per instance, register a shutdown hook to clean it
up when the JVM shuts down, so it can be reused more than once properly.

Result
------
Correct shutdown of the detector when we extend it in the future.

Thanks to @mp911de for the hint!

Change-Id: I78c6c269c1634c5d11d0d452325bb9677a0bc3ce
Reviewed-on: http://review.couchbase.org/63525
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>