JVMCBC-332: Implement VBucket Retry changes according to RFC.
Motivation
----------
In the existing suite of client libraries there is no
consistent behaviour defined in terms of how to perform
vbucket retries. This has resulted in a variety of methods
being employed, each with its own benefits and drawbacks.
We now need to align all the SDKs to have similar behaviour
to ensure that performance expectations do not vary based
on SDK.

Modifications
-------------
This changeset brings core-io "in line" with the RFC and
effectively changes the following:

 - Previously, NMVB responses were handled like every other
   reason to retry with the configured backoff strategy. Now,
   there is a fixed delay (configurable through a property) set
   to 100ms by default which will be used on consecutive retry
   attempts. Note that the first retry is scheduled immediately.

 - The KeyValue locator now takes the retryCount into account and
   if its > 0 and there is a FF-Map it will always use it to
   dispatch the KV operation. Previously the FF-Map was not taken
   into account and we'd just wait until the normal config has a
   good vbucket information.

Note that if there is no FF-Map available, the Locator will use
the current config as usual.

Result
------
The Java SDK behaves as specified in the RFC in terms of NMVB
retries, leading to a consistent and defined behavior.

Change-Id: I33086bbda68b353d41d8ec661f4e681721efe7ec
Reviewed-on: http://review.couchbase.org/64018
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>