JVMCBC-334: N1QL parsing of SELECT RAW
Motivation:
-----------
In some corner cases, where maximum throughput must be achieved with
N1QL, using a `SELECT RAW somefield` statement can give a boost. But
the current streaming parser relies on detecting JSON Objects in the
results.

Modifications:
--------------
The QueryHandler has been modified to detect if results is an array
made of JSON Objects or just scalar values. In the later case, this is
a SELECT RAW and new parsing logic is applied.

The case where results is empty is also covered (considered a classic
object case).

Added tests for a complex 8 results response and 1 result response.

Results:
--------
SELECT RAW is now supported by the QueryHandler.

Change-Id: I4eb170b7b405abff309a10857f3919b821bea840
Reviewed-on: http://review.couchbase.org/63821
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>