JVMCBC-255: Change CoreEnvironment#shutdown to blocking.
Motivation
----------
Deprecated since 2.2.0, the shutdown method is now changed from async
to sync, leading to more straightforward behavior and less bugs in the
field (like users forgetting to block on the call and therefore having
an unclean shutdown).

Modifications
-------------
In the process of moving shutdown to sync, a few other things had
to be changed as well:

 - The disconnectTimeout has been moved from the CouchbasEnvironment
   down to the CoreEnvironment since we need one on the now blocking
   shutdown method. The java-client will be modified to take this
   into account as well.
 - Because .timeout() can't be properly used, the Blocking utilities
   have been moved over from java-client as well and can now be removed
   from there if needed. Reason is that if .timeout() is chained in,
   there are thread still pending and the shutdown tests start to
   fail.

Result
------
Easier access to blocking environment shutdown and sane defaults.

Change-Id: Ic68c236d8917a71baa5b72082da6dcbafa2ee5dd
Reviewed-on: http://review.couchbase.org/63848
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>