JVMCBC-271: Remove queryEnabled on the CoreEnvironment
Motivation
----------
This environment flag has been introduced in 2.0 when N1QL was still
experimental and you had to run a separate instance to make it work.
Since this flag is not needed anymore in the 2.2 timeframe we deprecated
it and now taking it away with 2.3, since it is useless and confusing.

Modifications
-------------
Remove the queryEnabled and queryPort from the CoreEnvironment, its
implementation. It also fixes the tests and adapts the internal parts
which made use of it conditionally.

Result
------
No confusing and unneeded queryEnabled/queryPort anymore.

Change-Id: I1391d0ad3ace6d80a7e399a0bf50e9c6e37b57d5
Reviewed-on: http://review.couchbase.org/63850
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>