JVMCBC-333: Support passing in a custom KeyStore.
Motivation
----------
It has been requested a few times that especially in larger deployments,
sharing keystores as files is not practical and done through other means.

As a result to support these cases, it must be possible to pass in the
initialized keystore directly instead of the core initializing it from
a file on the users behalf.

Modifications
-------------
Extend the environment so that a KeyStore can be optionally passed in
as well. If this is the case it gets used in the SSLEngineFactory and
it is not looking for a file to load it from.

Result
------
More flexibility in larger deployments on how to initialize the
SSL KeyStore.

Change-Id: Ida4ff3abf1624b7a53db45cafe7c27969af23c55
Reviewed-on: http://review.couchbase.org/63853
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>