JVMCBC-336: Metrics host should not include port.
Motivation
----------
The host section of the latency metric is intended for proper grouping
and including the port defeats this purpose. Previously every host/port
combination got its own toplevel which makes the second level (the service
type) useless.

Modifications
-------------
Only the host address is stored as the host, stripping away the port and
allowing for proper grouping. Integration test modified to assert this.

Result
------
Correct grouping now in the emitted metrics on the event bus and in the
logs.

Change-Id: Ia50c314939540294afad5109965897008dddc6e5
Reviewed-on: http://review.couchbase.org/64109
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>