JVMCBC-337: Subdoc Multi Lookup needs to expose CAS.
Motivation
----------
Due to an oversight the multi lookup in subdoc currently does not
carry the CAS as part of the response message, even if it is returned
from the server.

Modifications
-------------
This changeset adds the cas value to the MultiLookupResponse and modifies
test cases to assert for it.

Result
------
The CAS can now be carried over in higher level implementations.

Change-Id: I402e596850d28641f46325048690bdb0f22061b1
Reviewed-on: http://review.couchbase.org/64589
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>