JVMCBC-338: N1QL parsing response without metrics
Motivation:
-----------
The query service now honors a request that says metrics should not be
included in the response. As a consequence, the QueryHandler should now
take that possibility into account.

Modifications:
--------------
The QueryHandler has been modified to work with a response where the
metrics section is omitted.

Added a "split at every position" test without metrics.

Results:
--------
Response without metrics is now supported by the QueryHandler.

Change-Id: Ic8a5720de0cff92af730c0ec90721840b0acb129
Reviewed-on: http://review.couchbase.org/64597
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>