JVMCBC-338: Fix parsing of results only, no errors/metrics
Motivation:
-----------
The previous change was incomplete in the sense that a response with
only results followed by status would still fail and timeout.

Modifications:
--------------
The QueryHandler has been modified to work with a response where the
metrics section is omitted, and there's only results (no errors/warn).

This involves detecting end of a results section despite not having a
new { after said section => rely also on lastChunk.

Added a "success 1" test without metrics.
Modified 1 test so that it uses only one DefaultLastHttpContent instead
of several in a row (as that would trigger lastChunk detection).

Results:
--------
Response without metrics is now supported by the QueryHandler.

Change-Id: I48dadf20bb22b13e9ae2a0b789b3265515744aab
Reviewed-on: http://review.couchbase.org/64600
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>