JVMCBC-343: WaitStrategy needs to be created for every CouchbaseCore
Motivation
----------
The change introduced in JVMCBC-335 leads to a performance regression
if more than one CouchbaseCore instance is used. In the java-client
this means that if more than one CouchbaseCluster is used and the
environment is shared there is, under high throughput, more
contention on the used WaitStrategy.

Modifications
-------------
This change introduces a WaitStrategyFactory so that every time
the Core needs to use it a new WaitStrategy is created instead
of sharing one per environment.

Note that this technically breaks the Environment API since it
exposes a Factory instead of the WaitStrategy, but since it
has been introduced only in 1.3.0 and its an advanced API
the breakage is considered minimal and it is needed to fix
a performance regression.

Result
------
Fixed performance regression and more correct handling of
custom wait strategies.

Change-Id: If954261fcdd53d5038b903569d743d0b02cd0881
Reviewed-on: http://review.couchbase.org/65049
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>