JVMCBC-344: Replace deprecated Disruptor APIs.
Motivation
----------
The code uses two methods that are deprecated and require minor
code changes to move them over to the new APIs.

Modifications
-------------
The exception handler now sits on a new method and also when
creating the Disruptor now only the ThreadFactory is passed in
instead of the ExecutorService since thats not needed.

Result
------
Code is on stable disruptor API.

Change-Id: I7e371f667c9d1261d19a82e288f080f2c75de210
Reviewed-on: http://review.couchbase.org/65185
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>