JVMCBC-342: Decrease environment counter on shutdown and emit event.
Motivation
----------
To be more correct the environment now decreses its instance counter
so that when properly shutdown and then created no warning is issued

Modifications
-------------
In addition to the counter change a new event type has been added
that is emitted on the event bus. Note that since the event is
emitted right at startup time the event can only be captured if
a custom (or shared) event bus is passed in, which is possible if
for example the event bus is a target for an external monitoring
app.

Result
------
More correct handling of shutdown environments and a new event
type that helps getting identified about this situation.

Change-Id: Id9b403da01fd45a7ebcea1c71b7d41fc76c09fed
Reviewed-on: http://review.couchbase.org/65186
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>