JVMCBC-345: Don't ignore ConnectException during bootstrap.
Motivation
----------
A ConnectException is triggered for example if a query node is
in the cluster but the service stopped, then port 8093 bootstrap
will fail of course. Right now the error is swallowed and the
node stays in a silent CONNECTING state, which is not enough
to surface the error.

Note that once up and running this is not an issue because
as it stands now the code will initiate a reconnect in any
case.

Modifications
-------------
The result now explicitly checks for the ConnectException and
handles it similarly to other errors that can happen during
the netty connect phase.

Result
------
Visible error what's going wrong during bootstrap against a
downed query node in the cluster.

Change-Id: I22fb5cbdd5af1dfa43fedb738a14fccba8311afe
Reviewed-on: http://review.couchbase.org/65673
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>