JVMCBC-347: Maintain internal bootstrap state in config provider.
Motivation
----------
The bootstrapped flag right now is only used to not let the seed
nodes be overridden when the provider is already bootstrapped.

It turns out that we never set the bootstrapped flag to true.

Modifications
-------------
When a bucket is opened successfully the flag is set. We never
need to set it back though since once bootstrapped we don't
go back to "not bootstrapped".

Result
------
seed nodes can't be overridden once at least a single bucket is
opened for the first time.

Change-Id: I0cbd8a6b26f81acdf2c945415c6342845a22ce6d
Reviewed-on: http://review.couchbase.org/65717
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Simon Baslé <simon@couchbase.com>