JVMCBC-350: Add generic RestApiRequest for 8091
Motivation:
-----------
The SDK has native APIs for some number of cluster management calls,
but for the rest of the cluster's REST API, a simplified client
could make sense. This can be done for port 8091 endpoints via a
dedicated message in the core, using ConfigHandler.

Modifications:
--------------
Added a RestApiRequest and RestApiResponse message that is handled
by the ConfigHandler, made to perform API calls on the data/config
port (8091).

Result:
-------
Generic REST API calls can be performed from the core.

Change-Id: I2d5e4af73b2a769452d0eb9873d97e7e1560fbd4
Reviewed-on: http://review.couchbase.org/66152
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>