JVMCBC-354: Getting all design documents needs data node
Motivation:
-----------
In a MDS configuration, when requesting all design documents in a
bucket, the request can go to a node that doesn't have the view service
enabled. In such a case the server answers with a "no_ddocs_service"
error. The request must be routed to a node with the VIEW service
enabled.

This leads to NullPointerException on the client side (because the
response does not correspond to the expected design documents payload).

Modifications:
--------------
Made ConfigHandler routing able to skip some nodes in the list, like
eg. ViewHandler.

Only skipping occurs when the request is a GetDesignDocuments one and
the checked node doesn't have the VIEW service enabled.

Results:
--------
GetDesignDocumentsRequest are correctly routed to VIEW enabled nodes
in a MDS configuration.

Change-Id: Iad4dd787184cdee05c0812548582004003d5cb83
Reviewed-on: http://review.couchbase.org/66413
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Simon Baslé <simon@couchbase.com>