JVMCBC-355: Add identifier to UnicastAutoReleaseSubject
Motivation:
-----------
When an UnicastAutoReleaseSubject throws an error due to multiple
subscribers coming in or a single subscriber coming in after the
data has been auto-released, it is usually hard to know the exact
source of the error. In order to improve on that, each subject could
be assigned an id that would at least help tracing back to the code
and the type of data.

Modifications:
--------------
Added id field to the subject's State, which is used in the
exception message and can be set through a setter on the subject.

Made use of that feature in all points where an UnicastAutoRelease
is used: DCPConnection, ViewHandler and QueryHandler.
QueryHandler can even trace the requestID / clientContextID that
is triggering the issue.

Added unit tests that validates the exception message in failure
scenarios.

Results:
--------
Errors linked to UnicastAutoReleaseSubject can be better diagnosed.

Change-Id: Ifa54a86506fa72a1cebc288892278ef440911857
Reviewed-on: http://review.couchbase.org/66449
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>